+++ tools/pnnx/src/pass_level5/fuse_scaled_dot_product_attention.cpp
@@ -19,7 +19,7 @@
 #include <math.h>
 #include <string.h>
 
-#include <torch/csrc/api/include/torch/torch.h>
+#include <torch/csrc/api/include/torch/version.h>
 
 namespace pnnx {
